particle.io OneBus transfer failures
Sorry, this is my first ever use of Github....Don't know exactly what I am doing.
See this thread for more detail:

https://community.particle.io/t/onewire-library-bug-proposed-fix-june-2019-affects-ds18b20-and-other-onewire-devices/50589

Here is the OneWire.cpp file with some changes that FIX the error, my changes can be found by searchcing for //JC
OneWireFIX.txt

The current particle.io implementation of the OneWire protocol has a unique problem: transfers on the OneWire are subject to being “interrupted” by higher level system functions/calls. When this occurs inside the critical window of a OneWire “bit transfer”, that transfer will fail and result in a CRC error.

Solutions should always be designed to deal with a low percentage of CRC errors, but the current particle.io implementation has a rate that is too high.

Particularly troublesome, this “bit transfer fail” seems to cause havoc for the OneWire “Search for OneWire devices” algorithm which is “bit transfer” intensive. Many applications use this OneWire function to “find” and enumerate devices attached to the bus. With this “interrupt issue”, it is very common to see this algorithm fail to find the addresses of all attached devices, especially when there are 3+ devices attached to the bus.

SOLUTION:
Protect the critical windows of the OneWire protocol by REALLY blocking all interrupts during those critical “bit transfer” times. Although the OneWire protocol DOES implement a disable/enable interrupt procedure around these windows, higher level system operations are not prevented making their way in.

For the following devices, there is a FIX that works. That FIX uses ATOMIC_BLOCKS around the critical OneWire code to protect the bit transfers. That FIX works for:

Photon, Argon (mesh disabled), Xenon (mesh disabled), Boron (mesh disabled).

You see the pattern….currently a solution for “mesh enabled” devices is being sought.